Fixed most of the bugs

+++ ex1/src/ex1/Obstacle.java
@@ -1,7 +1,7 @@
 package ex1;
 
 public class Obstacle extends WorldObject {
-    public Obstacle (int x, int y, double radius, World world) {
+    public Obstacle (double x, double y, double radius, World world) {
         super(x, y, radius, world);
     }
 }
